Minor changes to get TBB working with clang plus libc++.
This patch is largely lifted from this patch by jckarter (Joe Groff):
https://github.com/jckarter/tbb/commit/eb2e9e9a

My change to "include/tbb/compat/tuple" is cosmetically
different but serves the same purpose. The block comment
is jckarter's.

I'm not exactly sure what's going on with the _allocator types.
According to C++11 (20.6.9), the standard allocator<T> can
allocate any type, not just T:

    template <typename T>
    struct allocator {
        ...
        template <typename U, typename... Args>
        void construct(U *p, Args&&... args);
    };

but user-defined Allocators (such as might be passed as template
type parameters to STL containers) don't have to implement a
construct() member function at all, if they don't want to.
TBB's _allocator<T> types implemented the equivalent of

    template <typename T>
    struct TBB_allocator {
        ...
        template <typename... Args>
        void construct(T *p, Args&&... args);
    };

It looks as if libc++'s std::list<T> relies on the ability
of the passed-in Allocator<T> to construct objects of type
__list_node<T> as well as objects of type T. jckarter's
solution is to templatize TBB's _allocator<T>::construct()
the same way as C++11's std::allocator<T>::construct().
This strikes me as inelegant, but I haven't found anything
else that works.

std::copy_exception() appears to be a pre-standardization
feature; the C++11 equivalent is std::make_exception_ptr().

To build TBB with libc++ on Mac OS X, use one of the two
following command lines:

    make compiler=clang-libcxx
    make compiler=clang-libcxx cpp0x=1