Add support for ARM Linux.
I'm not sure what the deal is with -DDO_ITT_NOTIFY, but
it's unconditionally defined for Linux (even on PowerPC?!),
and it tries to use inline x86 assembly ("lock xchg") that
definitely doesn't work on ARM. So I've turned it off for
ARM Linux. I've left it on for all other non-x86 targets,
even though I don't think that makes sense. Apparently it's
not the default on __APPLE__ targets, because I didn't see
the same problem when compiling for iOS.

In Makefile.tbbmalloc, one test needs to link with -ldl
or else we get "undefined reference to 'dlsym'".