TBB 4.1 update 2
The download files for the 20130116 stable release. The
tbb41_20130116oss_{win,lin,osx} packages contain binaries for Windows*,
Linux*, and OS X* respectively, as well as TBB header files, HTML
documentation, and examples. To use TBB, download and untar/unzip the
required package; the tbb41_20130116oss directory will be created with
the package content. We recommend to set TBBROOT environment variable
to contain the full path to that directory. The tbb41_20130116oss_src
package contains the sources and makefiles to build TBB, and also headers,
docs, and examples. If used together with one of the binary packages, we
recommend to untar/unzip both into the same directory.